update alerts to sweet alerts in enrollments.

+++ scripts/enrollment.js
@@ -69,7 +69,7 @@ function addEnrollment() {
   })
   .then((response) => response.text())
   .then(responseText => {
-    alert(responseText);
+    Swal.fire("Success", `${responseText}`, "success");
     displayEnrollments(); // updates the table
   }).catch (error => {
     alert('console error.');
@@ -117,7 +117,7 @@ function editEnrollment(button) {
     })
     .then((response) => response.text())
     .then((responseText) => {
-      alert(responseText);
+      Swal.fire("Success", `${responseText}`, "success");
       displayEnrollments();
     });
 
@@ -126,6 +126,28 @@ function editEnrollment(button) {
     // Edit mode
     button.textContent = "Save";
     cells[0].focus();
+
+    const Toast = Swal.mixin({
+      toast: true,
+      position: 'top-end',
+      icon: "info",
+      showConfirmButton: false,
+      timer: 2000,
+      timerProgressBar: true,
+      customClass: {
+        popup: 'colored-toast'
+      },
+      didOpen: (toast) => {
+        toast.addEventListener('mouseenter', Swal.stopTimer);
+        toast.addEventListener('mouseleave', Swal.resumeTimer);
+      }
+    });
+
+    Toast.fire({
+      icon: 'info',
+      title: 'Edit Mode Activated'
+    });
+
     updatedRow.length = 0; // clear the array.
   }
 }
@@ -135,19 +157,28 @@ function deleteEnrollment(button) {
   const row = button.closest('tr');
   const cells = row.querySelectorAll('td');
   const enrollmentId = cells[0].innerHTML;
-  
-  fetch(enrollmentEndpoint, {
-    method: 'DELETE',
-    headers: {
-      "Content-type": "application/x-www-form-urlencoded",
-    },
-    body: `enrollment_id=${enrollmentId}`,
-  })
-  .then((response) => response.text())
-  .then((responseText) => {
-    alert(responseText);
-    displayEnrollments();
-  })
+
+  Swal.fire({
+    title: "Do you want to delete this data?",
+    icon: "question",
+    showCancelButton: true,
+    confirmButtonText: "Delete",
+  }).then((result) => {
+    if (result.isConfirmed) {
+        fetch(enrollmentEndpoint, {
+          method: 'DELETE',
+          headers: {
+            "Content-type": "application/x-www-form-urlencoded",
+          },
+          body: `enrollment_id=${enrollmentId}`,
+        })
+        .then((response) => response.text())
+        .then(() => {
+          Swal.fire("Deleted", "", "success");
+          displayEnrollments();
+      })
+    }
+  });
 }
 
 function searchEnrollment() {
